refactor(gui): extract throughput graph into graph.ts
